# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Testosteronium API — User Console + Admin Console
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeviceInfo {
  device_id: string;
  hostname: string;
  username: string;
  ip_address?: string;
  mac_address?: string;
  os_name: string;
  os_version?: string;
  connected: boolean;
}

export interface ClientTicket {
  id: number;
  deviceId: string;
  subject: string;
  description: string;
  category: string;
  priority: string;
  status: string;
  /** @nullable */
  assignedAgent?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  attachmentName?: string | null;
  /** @nullable */
  attachmentMime?: string | null;
  /** @nullable */
  attachmentSize?: number | null;
}

export interface TicketListResponse {
  device_id: string;
  tickets: ClientTicket[];
}

export interface ClientComment {
  id: number;
  ticketId: number;
  author: string;
  authorRole: string;
  body: string;
  createdAt: string;
}

export interface TicketDetailResponse {
  ticket: ClientTicket;
  comments: ClientComment[];
}

export interface CommentInput {
  body: string;
  author?: string;
}

export interface CommentResponse {
  comment: ClientComment;
  status: string;
}

export interface AssetSummary {
  tracked: number;
  manual: number;
  fetching: number;
}

export interface TrackedAsset {
  id: number;
  hostname: string;
  /** @nullable */
  username?: string | null;
  ip: string;
  /** @nullable */
  mac_address?: string | null;
  /** @nullable */
  serial?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  specs?: string | null;
  status: string;
  first_seen?: string;
  last_seen?: string;
  /** @nullable */
  enrolled_via_token?: string | null;
}

export interface ManualAsset {
  id: number;
  name: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  owner?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  serial?: string | null;
  /** @nullable */
  notes?: string | null;
  source: string;
  addedOn?: string;
}

export interface ManualAssetInput {
  name: string;
  type?: string;
  owner?: string;
  location?: string;
  serial?: string;
  notes?: string;
}

export interface ManualAssetUpdate {
  name?: string;
  type?: string;
  owner?: string;
  location?: string;
  serial?: string;
  notes?: string;
}

export interface FetchingAsset {
  id: number;
  ip: string;
  mac: string;
  /** @nullable */
  guess?: string | null;
  /** @nullable */
  os_guess?: string | null;
  /** @nullable */
  connection?: string | null;
  first_seen?: string;
  last_seen?: string;
}

export interface EnrollmentToken {
  token: string;
  /** @nullable */
  label?: string | null;
  created_at: string;
  expires_at: string;
  /** @nullable */
  used_at?: string | null;
  /** @nullable */
  used_by_mac?: string | null;
  status: string;
}

export interface EnrollmentTokenInput {
  label?: string;
  expires_hours: number;
}

export interface OkResponse {
  status: string;
  /** @nullable */
  id?: number | null;
}

export type ListTicketsParams = {
device_id?: string;
};

export type GetTicketParams = {
device_id?: string;
};

export type AddCommentParams = {
device_id?: string;
};